order_service/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
confluent-kafka==2.8.0
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
idna==3.10
psycopg2==2.9.10
requests==2.32.3
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.3.0
